Machine_Translation_Eval_app/backend/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
unbabel-comet
transquest
torch
transformers
python-multipart
xlrd
sacrebleu
bert-score
huggingface_hub